move modules tests into separate source set
