+++ etomica-modules/src/test/java/etomica/crystalviewer/BLCPrimitiveHexagonalLatticePlaneTest.java
@@ -2,16 +2,15 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package etomica.modules.junit.crystalviewer;
+package etomica.crystalviewer;
 
 import junit.framework.TestCase;
 import etomica.api.IAtom;
 import etomica.api.IAtomList;
 import etomica.api.IVectorMutable;
-import etomica.lattice.crystal.PrimitiveOrthorhombic;
+import etomica.lattice.crystal.PrimitiveHexagonal;
 
-
-public class BLCPrimitiveOrthorhombicLatticePlaneTest extends TestCase {
+public class BLCPrimitiveHexagonalLatticePlaneTest extends TestCase {
 
 	private final int DEFAULT_SIZE = 5;
 	private final int DEFAULT_MILLER[] = {0,1,0};
@@ -19,11 +18,11 @@ public class BLCPrimitiveOrthorhombicLatticePlaneTest extends TestCase {
 	
 	private String funcName = "";
 
-	private double epsilon = 1.0E-5;;
+	private double epsilon = 1.0E-5;
 
 	private LatticePlaneTestUtility lptu = null;
 
-	public BLCPrimitiveOrthorhombicLatticePlaneTest(String name) {
+	public BLCPrimitiveHexagonalLatticePlaneTest(String name) {
 		super(name);
 		funcName = name;
 	}
@@ -32,7 +31,7 @@ public class BLCPrimitiveOrthorhombicLatticePlaneTest extends TestCase {
 		super.setUp();
 		if (lptu == null) {
 			lptu = new LatticePlaneTestUtility();			
-	        lptu.createLatticeAndBox(lptu.ORTHORHOMBIC, DEFAULT_MILLER, DEFAULT_BOX);
+	        lptu.createLatticeAndBox(lptu.HEXAGONAL, DEFAULT_MILLER, DEFAULT_BOX);
 	        lptu.setDimensions(DEFAULT_SIZE);
 		}
 	}
@@ -58,11 +57,11 @@ public class BLCPrimitiveOrthorhombicLatticePlaneTest extends TestCase {
     	double plane = 1.0;
     	IAtomList leafList = null;
 
-        lptu.createLatticeAndBox(lptu.ORTHORHOMBIC, DEFAULT_MILLER, DEFAULT_BOX);
+        lptu.createLatticeAndBox(lptu.HEXAGONAL, DEFAULT_MILLER, DEFAULT_BOX);
 
-        ((PrimitiveOrthorhombic)lptu.getLattice().getPrimitive()).setSizeA(cubicSize);
-        ((PrimitiveOrthorhombic)lptu.getLattice().getPrimitive()).setSizeB(cubicSize);
-        ((PrimitiveOrthorhombic)lptu.getLattice().getPrimitive()).setSizeC(cubicSize);
+
+        ((PrimitiveHexagonal)lptu.getLattice().getPrimitive()).setSizeAB(cubicSize);
+        ((PrimitiveHexagonal)lptu.getLattice().getPrimitive()).setSizeC(cubicSize);
         lptu.setDimensions(DEFAULT_SIZE);
         lptu.setLatticePlanePosition(plane);
 
@@ -70,24 +69,23 @@ public class BLCPrimitiveOrthorhombicLatticePlaneTest extends TestCase {
         lptu.getLatticePlane().setPrimitive(lptu.getLattice().getPrimitive());
         double spacePos = lptu.getLatticePlaneSpacePosition();
 
+        //Only (and all) atoms where x = 0 should be in plane.
     	leafList = lptu.getBox().getLeafList();
 
     	try {
 		    for(idx = 0; idx < leafList.getAtomCount(); idx++) {
-			    IAtom a =  leafList.getAtom(idx);
+			    IAtom a = leafList.getAtom(idx);
                 if(a.getPosition().getX(1) >= spacePos-epsilon &&
                    a.getPosition().getX(1) <= spacePos+epsilon) {
-            	    assertTrue(lptu.getLatticePlane().inPlane(
-            	    		a.getPosition()));
+            	    assertTrue(lptu.getLatticePlane().inPlane(a.getPosition()));
                 }
                 else {
-            	    assertFalse(lptu.getLatticePlane().inPlane(
-            	    		a.getPosition()));
+            	    assertFalse(lptu.getLatticePlane().inPlane(a.getPosition()));
                 }
 		    }
 		}
         catch (junit.framework.AssertionFailedError e) {
-		    IAtom a =  leafList.getAtom(idx);
+		    IAtom a = leafList.getAtom(idx);
             if(a.getPosition().getX(1) >= spacePos-epsilon &&
                a.getPosition().getX(1) <= spacePos+epsilon) {
             	System.out.println(funcName + " -> Atom position : " + a.getPosition() +
@@ -97,34 +95,30 @@ public class BLCPrimitiveOrthorhombicLatticePlaneTest extends TestCase {
             	System.out.println(funcName + " ->Atom position : " + a.getPosition() +
             			" should not be in plane but is.");       	
             }
-         	fail();
+        	fail();
         }
 
     } // End testStandard()
 
     /*
      * Miller indices = 0, 1, 0
-     * size of cell (A) = 1.5
-     * size of cell (B) = 1.75
-     * size of cell (C) = 2.0
+     * size of cell (A, B) = 1.0
+     * size of cell (A, B) = 2.3
      * cells per side = 5
-     * plane = -2.0
+     * plane = -1.0
      */
-    public void testCellSizeAllDifferentIncrease() {
+    public void testCCellSizeIncrease() {
 
     	int idx = 0;
-    	double cubicSizeA = 1.5;
-    	double cubicSizeB = 1.75;
-    	double cubicSizeC = 2.0;
-    	double plane = -2.0;
+    	double cubicSizeAB = 1.0;
+    	double cubicSizeC = 2.3;
+    	double plane = -1.0;
     	IAtomList leafList = null;
 
-        lptu.createLatticeAndBox(lptu.ORTHORHOMBIC, DEFAULT_MILLER, DEFAULT_BOX);
+        lptu.createLatticeAndBox(lptu.HEXAGONAL, DEFAULT_MILLER, DEFAULT_BOX);
 
-        ((PrimitiveOrthorhombic)lptu.getLattice().getPrimitive()).setSizeA(cubicSizeA);
-        ((PrimitiveOrthorhombic)lptu.getLattice().getPrimitive()).setSizeB(cubicSizeB);
-        ((PrimitiveOrthorhombic)lptu.getLattice().getPrimitive()).setSizeC(cubicSizeC);
-
+        ((PrimitiveHexagonal)lptu.getLattice().getPrimitive()).setSizeAB(cubicSizeAB);
+        ((PrimitiveHexagonal)lptu.getLattice().getPrimitive()).setSizeC(cubicSizeC);
         lptu.setDimensions(DEFAULT_SIZE);
         lptu.setLatticePlanePosition(plane);
 
@@ -132,19 +126,17 @@ public class BLCPrimitiveOrthorhombicLatticePlaneTest extends TestCase {
         lptu.getLatticePlane().setPrimitive(lptu.getLattice().getPrimitive());
         double spacePos = lptu.getLatticePlaneSpacePosition();
 
-    	leafList = lptu.getBox().getLeafList();
+        leafList = lptu.getBox().getLeafList();
 
     	try {
 		    for(idx = 0; idx < leafList.getAtomCount(); idx++) {
 			    IAtom a =  leafList.getAtom(idx);
                 if(a.getPosition().getX(1) >= spacePos-epsilon &&
                    a.getPosition().getX(1) <= spacePos+epsilon) {
-            	    assertTrue(lptu.getLatticePlane().inPlane(
-            	    		a.getPosition()));
+            	    assertTrue(lptu.getLatticePlane().inPlane(a.getPosition()));
                 }
                 else {
-            	    assertFalse(lptu.getLatticePlane().inPlane(
-            	    		a.getPosition()));
+            	    assertFalse(lptu.getLatticePlane().inPlane(a.getPosition()));
                 }
 		    }
 		}
@@ -159,7 +151,7 @@ public class BLCPrimitiveOrthorhombicLatticePlaneTest extends TestCase {
             	System.out.println(funcName + " ->Atom position : " + a.getPosition() +
             			" should not be in plane but is.");       	
             }
-         	fail();
+        	fail();
         }
     } // End testABCellSizeIncrease()
 
@@ -167,7 +159,7 @@ public class BLCPrimitiveOrthorhombicLatticePlaneTest extends TestCase {
      * Miller indices = 1, 3, 2
      * size of cell (A, B, C) = 1.0
      * cells per side = 9
-     * plane = 8.0
+     * plane = 7.0
      */
     public void testOddMillerIndicesDistantPlane() {
 
@@ -175,46 +167,47 @@ public class BLCPrimitiveOrthorhombicLatticePlaneTest extends TestCase {
     	double cubicSize = 1.0;
     	IAtomList leafList = null;
     	int size = 9;
-    	double plane = 8.0;
     	int itemsFound = 0;
+    	double plane = 7.0;
     	int[] millerIndices = new int[] { 1, 3, 2 };
         double actualPlane[][] =
-                { { -4.0, 2.0, 3.0 }, { -4.0, 4.0, 0.0 }, { -3.0, 1.0, 4.0},
-                  { -3.0, 3.0, 1.0 }, { -2.0, 2.0, 2.0 }, { -2.0, 4.0, -1.0},
-                  { -1.0, 1.0, 3.0 }, { -1.0, 3.0, 0.0 }, { 0.0, 0.0, 4.0},
-                  { 0.0, 2.0, 1.0 }, { 0.0, 4.0, -2.0 }, { 1.0, 1.0, 2.0},
-                  { 1.0, 3.0, -1.0 }, { 2.0, 0.0, 3.0 }, { 2.0, 2.0, 0.0},
-                  { 2.0, 4.0, -3.0 }, { 3.0, -1.0, 4.0 }, { 3.0, 1.0, 1.0},
-                  { 3.0, 3.0, -2.0 }, { 4.0, 0.0, 2.0 }, { 4.0, 2.0, -1.0},
-                  { 4.0, 4.0, -4.0} };
+         { { -4.5, 0.8660254037844388, 4.0 },{ -5.5, 2.598076211353317, 1.0 },
+           { -4.0, 1.7320508075688772, 2.0 },{ -5.0, 3.4641016151377557, -1.0 },
+           { -2.5, 0.866025403784439, 3.0 },{ -3.5, 2.5980762113533173, 0.0 },
+           { -1.0, 0.0, 4.0 },{ -2.0, 1.7320508075688772, 1.0 },
+           { -3.0, 3.4641016151377557, -2.0 },{ -0.5, 0.8660254037844388, 2.0 },
+           { -1.5, 2.598076211353317, -1.0 },{ 1.0, 0.0, 3.0 },
+           { 0.0, 1.7320508075688772, 0.0 },{ -1.0, 3.4641016151377557, -3.0 },
+           { 2.5, -0.8660254037844393, 4.0 },{ 1.5, 0.8660254037844388, 1.0 },
+           { 0.5, 2.5980762113533173, -2.0 },{ 3.0, 0.0, 2.0 },
+           { 2.0, 1.7320508075688772, -1.0 },{ 1.0, 3.4641016151377557, -4.0 },
+           { 4.5, -0.8660254037844393, 3.0 },{ 3.5, 0.8660254037844388, 0.0 },
+           { 2.5, 2.5980762113533173, -3.0 } };
 
         DoubleTwoDArray dd = new DoubleTwoDArray(actualPlane);
 
-        lptu.createLatticeAndBox(lptu.ORTHORHOMBIC, millerIndices, new int[] {size, size, size});
-        
-        ((PrimitiveOrthorhombic)lptu.getLattice().getPrimitive()).setSizeA(cubicSize);
-        ((PrimitiveOrthorhombic)lptu.getLattice().getPrimitive()).setSizeB(cubicSize);
-        ((PrimitiveOrthorhombic)lptu.getLattice().getPrimitive()).setSizeC(cubicSize);
+        lptu.createLatticeAndBox(lptu.HEXAGONAL, millerIndices, new int[] {size, size, size});
+
+        ((PrimitiveHexagonal)lptu.getLattice().getPrimitive()).setSizeAB(cubicSize);
+        ((PrimitiveHexagonal)lptu.getLattice().getPrimitive()).setSizeC(cubicSize);
         lptu.setDimensions(size);
         lptu.setLatticePlanePosition(plane);
 
         // This needs to come after lattice changes
         lptu.getLatticePlane().setPrimitive(lptu.getLattice().getPrimitive());
 
-    	leafList = lptu.getBox().getLeafList();
+        leafList = lptu.getBox().getLeafList();
 
     	try {
 		    for(idx = 0; idx < leafList.getAtomCount(); idx++) {
 			    IAtom a = leafList.getAtom(idx);
 
 			    if(dd.contains(makeArray(a.getPosition())) == true) {
-	            	itemsFound++;
-            	    assertTrue(lptu.getLatticePlane().inPlane(
-            	    		a.getPosition()));
+			    	itemsFound++;
+            	    assertTrue(lptu.getLatticePlane().inPlane(a.getPosition()));
                 }
                 else {
-            	    assertFalse(lptu.getLatticePlane().inPlane(
-            	    		a.getPosition()));
+            	    assertFalse(lptu.getLatticePlane().inPlane(a.getPosition()));
                 }
 		    }
 		}
@@ -228,12 +221,12 @@ public class BLCPrimitiveOrthorhombicLatticePlaneTest extends TestCase {
             	System.out.println(funcName + " ->Atom position : " + a.getPosition() +
             			" should not be in plane but is.");       	
             }
-         	fail();
+        	fail();
         }
 
         assertEquals(actualPlane.length, itemsFound);
 
-    } // End testOddMillerIndicesDistantPlane()
+    } // End testCellSizeIncrease()
 
     /*
      * Miller indices = 0, 1, 0
@@ -249,12 +242,11 @@ public class BLCPrimitiveOrthorhombicLatticePlaneTest extends TestCase {
     	int dimensionSize = 4;
     	double plane = 0.0;
 
-        lptu.createLatticeAndBox(lptu.ORTHORHOMBIC, DEFAULT_MILLER,
+        lptu.createLatticeAndBox(lptu.HEXAGONAL, DEFAULT_MILLER,
         		                   new int[] {dimensionSize, dimensionSize, dimensionSize});
 
-        ((PrimitiveOrthorhombic)lptu.getLattice().getPrimitive()).setSizeA(cubicSize);
-        ((PrimitiveOrthorhombic)lptu.getLattice().getPrimitive()).setSizeB(cubicSize);
-        ((PrimitiveOrthorhombic)lptu.getLattice().getPrimitive()).setSizeC(cubicSize);
+        ((PrimitiveHexagonal)lptu.getLattice().getPrimitive()).setSizeAB(cubicSize);
+        ((PrimitiveHexagonal)lptu.getLattice().getPrimitive()).setSizeC(cubicSize);
         lptu.setDimensions(dimensionSize);
         lptu.setLatticePlanePosition(plane);
 
@@ -266,15 +258,14 @@ public class BLCPrimitiveOrthorhombicLatticePlaneTest extends TestCase {
     	try {
 		    for(idx = 0; idx < leafList.getAtomCount(); idx++) {
 			    IAtom a =  leafList.getAtom(idx);
-            	assertFalse(lptu.getLatticePlane().inPlane(
-            	    	a.getPosition()));
+            	assertFalse(lptu.getLatticePlane().inPlane(a.getPosition()));
 		    }
 		}
         catch (junit.framework.AssertionFailedError e) {
 		    IAtom a =  leafList.getAtom(idx);
             System.out.println(funcName + " ->Atom position : " + a.getPosition() +
-            			" should not be in plane but is.");
-             fail();
+            			" should not be in plane but is.");  
+            fail();
         }
     	
     } // End testEvenAtomsPerSideZeroPlane
@@ -293,12 +284,11 @@ public class BLCPrimitiveOrthorhombicLatticePlaneTest extends TestCase {
     	int dimensionSize = 4;
     	double plane = 0.5;
 
-        lptu.createLatticeAndBox(lptu.ORTHORHOMBIC, DEFAULT_MILLER,
+        lptu.createLatticeAndBox(lptu.HEXAGONAL, DEFAULT_MILLER,
         		                   new int[] {dimensionSize, dimensionSize, dimensionSize});
 
-        ((PrimitiveOrthorhombic)lptu.getLattice().getPrimitive()).setSizeA(cubicSize);
-        ((PrimitiveOrthorhombic)lptu.getLattice().getPrimitive()).setSizeB(cubicSize);
-        ((PrimitiveOrthorhombic)lptu.getLattice().getPrimitive()).setSizeC(cubicSize);
+        ((PrimitiveHexagonal)lptu.getLattice().getPrimitive()).setSizeAB(cubicSize);
+        ((PrimitiveHexagonal)lptu.getLattice().getPrimitive()).setSizeC(cubicSize);
         lptu.setDimensions(dimensionSize);
         lptu.setLatticePlanePosition(plane);
 
@@ -314,12 +304,10 @@ public class BLCPrimitiveOrthorhombicLatticePlaneTest extends TestCase {
 
                 if(a.getPosition().getX(1) >= spacePos-epsilon &&
                    a.getPosition().getX(1) <= spacePos+epsilon) {
-            	    assertTrue(lptu.getLatticePlane().inPlane(
-            	    		a.getPosition()));
+            	    assertTrue(lptu.getLatticePlane().inPlane(a.getPosition()));
                 }
                 else {
-            	    assertFalse(lptu.getLatticePlane().inPlane(
-            	    		a.getPosition()));
+            	    assertFalse(lptu.getLatticePlane().inPlane(a.getPosition()));
                 }
 		    }
 		}
@@ -334,117 +322,105 @@ public class BLCPrimitiveOrthorhombicLatticePlaneTest extends TestCase {
                  System.out.println(funcName + " ->Atom position : " + a.getPosition() +
                  			" should not be in plane but is.");       	
              }
-         	fail();
+        	fail();
         }
     	
     } // End testEvenAtomsPerSideZeroPt5Plane
 
     /*
-     * Miller indices = 2, 2, 1
-     * size of cell (A) = 0.7
-     * size of cell (B) = 0.8
-     * size of cell (C) = 0.9
-     * cells per side = 7
-     * plane = 2.95
+     * Miller indices = 1, 1, 2
+     * size of cell (A, B) = 1.4
+     * size of cell (C) = 1.4
+     * cells per side = 5
+     * plane = 0.05
      */
-    public void testPlaneMinusFiveHundreths() {
+    public void testPlaneMinusFiveHundreth() {
 
     	int idx = 0;
-    	double cubicSizeA = 0.7;
-    	double cubicSizeB = 0.8;
-    	double cubicSizeC = 0.9;
+    	double cubicSizeAB = 1.4;
+    	double cubicSizeC = 1.4;
     	IAtomList leafList = null;
-    	int size = 7;
-    	double plane = 2.95;
-    	int[] millerIndices = new int[] { 2, 2, 1 };
+    	int size = 5;
+    	double plane = 0.05;
+    	int[] millerIndices = new int[] { 1, 1, 2 };
 
-        lptu.createLatticeAndBox(lptu.ORTHORHOMBIC, millerIndices, new int[] {size, size, size});
-        
-        ((PrimitiveOrthorhombic)lptu.getLattice().getPrimitive()).setSizeA(cubicSizeA);
-        ((PrimitiveOrthorhombic)lptu.getLattice().getPrimitive()).setSizeB(cubicSizeB);
-        ((PrimitiveOrthorhombic)lptu.getLattice().getPrimitive()).setSizeC(cubicSizeC);
+        lptu.createLatticeAndBox(lptu.HEXAGONAL, millerIndices, new int[] {size, size, size});
+
+        ((PrimitiveHexagonal)lptu.getLattice().getPrimitive()).setSizeAB(cubicSizeAB);
+        ((PrimitiveHexagonal)lptu.getLattice().getPrimitive()).setSizeC(cubicSizeC);
         lptu.setDimensions(size);
         lptu.setLatticePlanePosition(plane);
 
         // This needs to come after lattice changes
         lptu.getLatticePlane().setPrimitive(lptu.getLattice().getPrimitive());
 
-    	leafList = lptu.getBox().getLeafList();
+        leafList = lptu.getBox().getLeafList();
 
     	try {
 		    for(idx = 0; idx < leafList.getAtomCount(); idx++) {
 			    IAtom a = leafList.getAtom(idx);
 
-            	assertFalse(lptu.getLatticePlane().inPlane(
-            	    		a.getPosition()));
+            	assertFalse(lptu.getLatticePlane().inPlane(a.getPosition()));
 		    }
 		}
         catch (junit.framework.AssertionFailedError e) {
 		    IAtom a =  leafList.getAtom(idx);
             System.out.println(funcName + " ->Atom position : " + a.getPosition() +
-            			" should be in plane but is not.");
-         	fail();
+            			" should not be in plane but is.");
+        	fail();
         }
 
-
-    } // End testPlaneMinusFiveHundreths()
+    } // End testPlaneMinusFiveHundreth()
 
     /*
-     * Miller indices = 2, 2, 1
-     * size of cell (A) = 0.7
-     * size of cell (B) = 0.8
-     * size of cell (C) = 0.9
-     * cells per side = 7
-     * plane = 3.0
+     * Miller indices = 1, 1, 2
+     * size of cell (A, B) = 1.4
+     * size of cell (C) = 1.4
+     * cells per side = 5
+     * plane = 0.0
      */
     public void testPlane() {
 
     	int idx = 0;
-    	double cubicSizeA = 0.7;
-    	double cubicSizeB = 0.8;
-    	double cubicSizeC = 0.9;
+    	double cubicSizeAB = 1.4;
+    	double cubicSizeC = 1.4;
     	IAtomList leafList = null;
-    	int size = 7;
-    	double plane = 3.0;
+    	int size = 5;
     	int itemsFound = 0;
-    	int[] millerIndices = new int[] { 2, 2, 1 };
+    	double plane = 0.0;
+    	int[] millerIndices = new int[] { 1, 1, 2 };
         double actualPlane[][] =
-                  { { -2.1, 2.4, 2.7 }, { -1.4, 1.6, 2.7 }, { -1.4, 2.4, 0.9 },
-                    { -0.7, 0.8, 2.7 }, { -0.7, 1.6, 0.9 }, { -0.7, 2.4, -0.9 },
-                    { 0.0, 0.0, 2.7}, { 0.0, 0.8, 0.9 }, { 0.0, 1.6, -0.9 },
-                    { 0.0, 2.4, -2.7 }, { 0.7, -0.8, 2.7 }, { 0.7, 0.0, 0.9 },
-                    { 0.7, 0.8, -0.9 }, { 0.7, 1.6, -2.7 }, { 1.4, -1.6, 2.7 },
-                    { 1.4, -0.8, 0.9 }, { 1.4, 0.0, -0.9 }, { 1.4, 0.8, -2.7 },
-                    { 2.1, -2.4, 2.7 }, { 2.1, -1.6, 0.9 }, { 2.1, -0.8, -0.9 },
-                    { 2.1, 0.0, -2.7} };
-
+                { { -1.4, -2.4248711305964283, 2.8 }, { -2.8, 0.0, 1.4 },
+                  { -4.2, 2.4248711305964283, 0.0 }, { -0.7, -1.2124355652982142, 1.4 },
+                  { -2.1, 1.2124355652982142, 0.0 }, { 1.4, -2.4248711305964283, 1.4 },
+                  { 0.0, 0.0, 0.0 }, { -1.4, 2.4248711305964283, -1.4 },
+                  { 2.1, -1.2124355652982142, 0.0 }, { 0.7, 1.2124355652982142, -1.4 },
+                  { 4.2, -2.4248711305964283, 0.0 }, { 2.8, 0.0, -1.4 },
+                  { 1.4, 2.4248711305964283, -2.8 } };
         DoubleTwoDArray dd = new DoubleTwoDArray(actualPlane);
 
-        lptu.createLatticeAndBox(lptu.ORTHORHOMBIC, millerIndices, new int[] {size, size, size});
-        
-        ((PrimitiveOrthorhombic)lptu.getLattice().getPrimitive()).setSizeA(cubicSizeA);
-        ((PrimitiveOrthorhombic)lptu.getLattice().getPrimitive()).setSizeB(cubicSizeB);
-        ((PrimitiveOrthorhombic)lptu.getLattice().getPrimitive()).setSizeC(cubicSizeC);
+        lptu.createLatticeAndBox(lptu.HEXAGONAL, millerIndices, new int[] {size, size, size});
+
+        ((PrimitiveHexagonal)lptu.getLattice().getPrimitive()).setSizeAB(cubicSizeAB);
+        ((PrimitiveHexagonal)lptu.getLattice().getPrimitive()).setSizeC(cubicSizeC);
         lptu.setDimensions(size);
         lptu.setLatticePlanePosition(plane);
 
         // This needs to come after lattice changes
         lptu.getLatticePlane().setPrimitive(lptu.getLattice().getPrimitive());
 
-    	leafList = lptu.getBox().getLeafList();
+        leafList = lptu.getBox().getLeafList();
 
     	try {
 		    for(idx = 0; idx < leafList.getAtomCount(); idx++) {
 			    IAtom a = leafList.getAtom(idx);
 
 			    if(dd.contains(makeArray(a.getPosition())) == true) {
-	            	itemsFound++;
-            	    assertTrue(lptu.getLatticePlane().inPlane(
-            	    		a.getPosition()));
+			    	itemsFound++;
+            	    assertTrue(lptu.getLatticePlane().inPlane(a.getPosition()));
                 }
                 else {
-            	    assertFalse(lptu.getLatticePlane().inPlane(
-            	    		a.getPosition()));
+            	    assertFalse(lptu.getLatticePlane().inPlane(a.getPosition()));
                 }
 		    }
 		}
@@ -458,7 +434,7 @@ public class BLCPrimitiveOrthorhombicLatticePlaneTest extends TestCase {
             	System.out.println(funcName + " ->Atom position : " + a.getPosition() +
             			" should not be in plane but is.");       	
             }
-         	fail();
+        	fail();
         }
 
         assertEquals(actualPlane.length, itemsFound);
@@ -466,56 +442,51 @@ public class BLCPrimitiveOrthorhombicLatticePlaneTest extends TestCase {
     } // End testPlane()
 
     /*
-     * Miller indices = 2, 2, 1
-     * size of cell (A) = 0.7
-     * size of cell (B) = 0.8
-     * size of cell (C) = 0.9
-     * cells per side = 7
-     * plane = 3.05
+     * Miller indices = 1, 1, 2
+     * size of cell (A, B) = 1.4
+     * size of cell (C) = 1.4
+     * cells per side = 5
+     * plane = 0.05
      */
-    public void testPlanePlusFiveHundreths() {
+    public void testPlanePlusFiveHundreth() {
 
     	int idx = 0;
-    	double cubicSizeA = 0.7;
-    	double cubicSizeB = 0.8;
-    	double cubicSizeC = 0.9;
+    	double cubicSizeAB = 1.4;
+    	double cubicSizeC = 1.4;
     	IAtomList leafList = null;
-    	int size = 7;
-    	double plane = 3.05;
-    	int[] millerIndices = new int[] { 2, 2, 1 };
+    	int size = 5;
+    	double plane = 0.05;
+    	int[] millerIndices = new int[] { 1, 1, 2 };
 
-        lptu.createLatticeAndBox(lptu.ORTHORHOMBIC, millerIndices, new int[] {size, size, size});
-        
-        ((PrimitiveOrthorhombic)lptu.getLattice().getPrimitive()).setSizeA(cubicSizeA);
-        ((PrimitiveOrthorhombic)lptu.getLattice().getPrimitive()).setSizeB(cubicSizeB);
-        ((PrimitiveOrthorhombic)lptu.getLattice().getPrimitive()).setSizeC(cubicSizeC);
+        lptu.createLatticeAndBox(lptu.HEXAGONAL, millerIndices, new int[] {size, size, size});
+
+        ((PrimitiveHexagonal)lptu.getLattice().getPrimitive()).setSizeAB(cubicSizeAB);
+        ((PrimitiveHexagonal)lptu.getLattice().getPrimitive()).setSizeC(cubicSizeC);
         lptu.setDimensions(size);
         lptu.setLatticePlanePosition(plane);
 
         // This needs to come after lattice changes
         lptu.getLatticePlane().setPrimitive(lptu.getLattice().getPrimitive());
 
-    	leafList = lptu.getBox().getLeafList();
+        leafList = lptu.getBox().getLeafList();
 
     	try {
 		    for(idx = 0; idx < leafList.getAtomCount(); idx++) {
 			    IAtom a = leafList.getAtom(idx);
 
-            	assertFalse(lptu.getLatticePlane().inPlane(
-            	    		a.getPosition()));
+            	assertFalse(lptu.getLatticePlane().inPlane(a.getPosition()));
 		    }
 		}
         catch (junit.framework.AssertionFailedError e) {
 		    IAtom a =  leafList.getAtom(idx);
             System.out.println(funcName + " ->Atom position : " + a.getPosition() +
-            			" should be in plane but is not.");
-         	fail();
+            			" should not be in plane but is.");
+        	fail();
         }
 
+    } // End testPlanePlusFiveHundreth()
 
-    } // End testPlanePlusFiveHundreths()
-
-    public class DoubleTwoDArray {
+    public static class DoubleTwoDArray {
     	private double[][] array;
     	private double epsilon = 1.0E-5;
