+++ etomica-modules/src/test/java/etomica/crystalviewer/BLCPrimitiveMonoclinicLatticePlaneTest.java
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package etomica.modules.junit.crystalviewer;
+package etomica.crystalviewer;
 
 import junit.framework.TestCase;
 import etomica.api.IAtom;
